requirements.txt
streamlit
google-generativeai
huggingface_hub
reportlab
Pillow
requests
